refactor: Square의 공통된 부분을 추상 클래스로 분리
