+++ src/main/java/chess/model/board/AbstractSquare.java
@@ -1,0 +1,22 @@
+package chess.model.board;
+
+import chess.model.piece.Piece;
+
+public abstract class AbstractSquare implements Square{
+
+    private final Position position;
+
+    public AbstractSquare(final Position position) {
+        this.position = position;
+    }
+
+    @Override
+    public Square receivePiece(final Piece piece) {
+        return new PieceSquare(this.position, piece);
+    }
+
+    @Override
+    public Square removePiece() {
+        return new EmptySquare(position);
+    }
+}
