+++ src/main/java/chess/model/board/EmptySquare.java
@@ -4,27 +4,15 @@ import chess.model.Color;
 import chess.model.Type;
 import chess.model.piece.Piece;
 
-public class EmptySquare implements Square {
-
-    private final Position position;
+public class EmptySquare extends AbstractSquare {
 
     public EmptySquare(final Position position) {
-        this.position = position;
-    }
-
-    @Override
-    public Square receivePiece(final Piece piece) {
-        return new PieceSquare(this.position, piece);
+        super(position);
     }
 
     @Override
     public Square movePiece(final Position position) {
         throw new UnsupportedOperationException("지원하지 않는 기능입니다.");
-    }
-
-    @Override
-    public Square removePiece() {
-        return new EmptySquare(position);
     }
 
     @Override
